// frontend/src/gen/DriveShare.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DriveRecipient } from "./DriveRecipient";

/**
 * The members who were granted reader access.
 */
export type DriveShare = { fileId: string, recipients: Array<DriveRecipient>, };